+++ dev/documents/CHANGELOG.md
@@ -1,9 +1,13 @@
 # ChangeLog
+## Known BUGs
+- Compositor Node RenderLayers cant be set correctly because it's linked to the scene output settings.
+- Temperature and Tint of ColorManagedViewSettings cant be set correctly.
+
 ## [1.0.5] - 2025-08-
 ### Fixed
-- Compositor node `File Output` can not be set.
+- Compositor node `File Output` can not be set correctly.
 
-## [1.0.4] - 2025-08-
+## [1.0.4] - 2025-08-14
 ### Fixed
 - Single node with "/", "\" in it's name can't be saved.
 - `NodeGroup` can not be set correctly in Blender 4.2.